Push all asynchronous activity back to the main thread.
The scheduler now accepts, in addition timers, also "tasks" for
immediate execution at the next opportunity. If any of our threads
needs to do work, we now schedule that back to the scheduler as such a
task. It will then be processed as part of the main loop. This means
that thread-safety becomes much less of an issue, and hence require
less locking. We'll clean lock up in a subsequent thread.