Switch from macOS installer to system extensions.
I believe this is finally the right model on macOS: zeek-agent now
runs as a system extension that gets installed by an outer container
app. The outer app does three things: (1) it interacts with the user
through a small UI (not very pretty yet), including offering some
basic configuration; (2) it asks the system to install the extension
persistently (it'll keep running even when the app terminates); and
(3) it communicates with the running extension through XPC for
configuration and status checks.

This also already prepares for adding a second, network extension
later: the infrastructure is mostly there, but no code yet.