Add a macOS network extension to the agent.
On macOS, the agent was already running as a system extension. This
adds the appropriate magic to hook into the OS as a "Network
Extension", so that we gain access to network activity. We don't use
this yet for anything, but this gets all the infrastructure pieces in
place (which turned out to be quite tricky ...).

A specific change here concerns moving the agent's main logic into a
separate thread on macOS, because a Network Extension (apparently)
needs to take over the main thread. That comes with some required
reorganization of the logic for option parsing so that our
macOS-specific code for retrieving system `defaults` doesn't execute
before we've moved on to the new thread.