Rework handling of macOS deployment target.
By default, our deployment target is now the current platform's major
version. That avoids problems with system libraries not matching in
for what they were built.

To set a different deployment target, there's now a corresponding
`configure` option. We set that for CI builds.

As part of this change, we also raise the CI's deployment target to
from Big Sur to Monterey.

Note: The CI changes are untested at this point.